[restructure] folder name
